Maecss 23 integrate ap is with all built u is

+++ user/auth/auth.js
@@ -24,7 +24,7 @@ $(function () {
   }
   const currentUser = userService.getCurrentUser();
 
-  if (currentUser.role !== "customer") {
+  if (currentUser && currentUser.role !== "customer") {
     userService.deleteCurrentUser();
     location.reload();
   }
